Add files via upload
it's a school administration tool. This program defines two classes, Student and School, which represent a student and a school respectively. The School class has methods for adding and removing students, getting a list of student names, and getting the average grade of all students.

The program's main loop displays a menu of options for the user to choose from. Depending on the user's choice, the program either adds a new student, removes an existing student, displays a list of student names, displays the average grade of all students, or quits the program.