Add files via upload
This code is used for web scraping programs. This code sends a GET request to "https://www.example.com", retrieves the HTML content, and then uses BeautifulSoup to parse the content and extract the text of the first <h1> tag. You can modify this code to scrape any website by changing the URL and using different BeautifulSoup methods to find and extract the data you need. However, please note that web scraping may be subject to legal and ethical considerations